# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ZeeoWork API Name",
        "description": "ZeeoWork API Description",
        "contact": {
            "email": "tech@zeework.ir"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://devpandora.zeework.ir/"
        }
    ],
    "paths": {
...
        "/api/hr/dashboard": {
            "get": {
                "tags": [
                    "OrganizationAdmin/Dashboard"
                ],
                "summary": "Get dashboard data for the organization admin",
                "operationId": "50e1386f0e624bc93d1eea1b34bcf474",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "participation_rate": {
                                                    "type": "object",
                                                    "example": {
                                                        "percentage": 75,
                                                        "description": "نرخ مشارکت در ارزیابی‌ها"
                                                    }
                                                },
                                                "health_scores_summary": {
                                                    "type": "object",
                                                    "example": {
                                                        "hq_score": {
                                                            "value": 85,
                                                            "label": "امتیاز کلی HQ"
                                                        }
                                                    }
                                                },
                                                "teams_status": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "object"
                                                    },
                                                    "example": [
                                                        {
                                                            "team_name": "تیم فنی",
                                                            "hq_score": 92,
                                                            "status": "عالی"
                                                        }
                                                    ]
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/hr/departments": {
            "get": {
                "tags": [
                    "OrganizationAdmin/Departments"
                ],
                "summary": "OA docs for this route",
                "operationId": "2783a7197e73552f9cc42ccace08a28f",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Limit",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/OrganizationAdminDepartmentResource"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OrganizationAdmin/Departments"
                ],
                "summary": "OA docs for this route",
                "operationId": "3175ebe05683b879ab03e1527095c7f4",
                "requestBody": {
                    "$ref": "#/components/schemas/OrganizationAdminDepartmentRequest"
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hr/departments/{id}": {
            "get": {
                "tags": [
                    "OrganizationAdmin/Departments"
                ],
                "summary": "OA docs for this route",
                "operationId": "b3e62e4e160849e44a89306245f05b44",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationAdmin ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/OrganizationAdminDepartmentResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "OrganizationAdmin/Departments"
                ],
                "summary": "OA docs for this route",
                "operationId": "c5789911869305f1a6846b0106642438",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationAdmin ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "$ref": "#/components/schemas/OrganizationAdminDepartmentRequest"
                },
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "OrganizationAdmin/Departments"
                ],
                "summary": "OA docs for this route",
                "operationId": "e0f14532fe3ad7572577285761e59f56",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "OrganizationAdmin ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "خروجی استاندارد",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ApiResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/hr/departments/list-for-select": {
            "get": {
                "tags": [
                    "OrganizationAdmin/Departments"
                ],
                "summary": "Get a simple list of departments for dropdowns",
                "operationId": "a3dce1ef1193316ef9efbcdd618e2e6e",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ApiResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "messages": {
                        "description": "آرایه‌ای از پیام‌ها برای کاربر (می‌تواند شامل پیام‌های موفقیت، خطا یا اطلاعاتی باشد)",
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "عملیات با موفقیت انجام شد."
                        },
                        "example": [
                            "عملیات موفقیت‌آمیز بود",
                            "پیام دوم در صورت وجود"
                        ],
                        "nullable": true
                    },
                    "data": {
                        "type": "object",
                        "example": {
                            "id": 1,
                            "name": "Ali"
                        },
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "OrganizationAdminDepartmentRequest": {
                "title": "Organization Admin Department Request",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "description": "نام دپارتمان (در هر سازمان یکتا)",
                        "type": "string",
                        "maxLength": 255,
                        "example": "منابع انسانی"
                    },
                    "description": {
                        "description": "توضیحات دپارتمان (اختیاری)",
                        "type": "string",
                        "maxLength": 1000,
                        "example": "این دپارتمان مسئول جذب و مدیریت منابع انسانی است.",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "OrganizationAdminDepartmentResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 3
                    },
                    "name": {
                        "type": "string",
                        "example": "واحد مالی"
                    },
                    "description": {
                        "type": "string",
                        "example": "این دپارتمان مسئول امور مالی شرکت است.",
                        "nullable": true
                    },
                    "employee_count": {
                        "description": "تعداد کارمندان دپارتمان (در صورت بارگذاری relation)",
                        "type": "integer",
                        "example": 12,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-01 08:30:00"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-06-10 11:21:00"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}